codegen: delete output dir only if it exists

+++ src/Odachi.CodeGen/CodeGenerator.cs
@@ -94,7 +94,7 @@ namespace Odachi.CodeGen
 
 			var packageContext = CreatePackageContext(package, options);
 
-			if (options.CleanOutputPath)
+			if (options.CleanOutputPath && Directory.Exists(packageContext.Path))
 			{
 				Directory.Delete(packageContext.Path, true);
 			}
